# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * CropGuard AI API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DemoLoginInput {
  name: string;
  contact: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  farmer: 'farmer',
  officer: 'officer',
  kvk: 'kvk',
} as const;

export interface User {
  id: string;
  name: string;
  role: UserRole;
  village: string;
  state: string;
}

export interface AuthSession {
  token: string;
  user: User;
}

export interface DetectionInput {
  sampleId?: string;
  /** @nullable */
  imageDataUrl?: string | null;
  crop: string;
  /** @nullable */
  fieldId?: number | null;
}

export type DetectionType = typeof DetectionType[keyof typeof DetectionType];


export const DetectionType = {
  disease: 'disease',
  pest: 'pest',
  healthy: 'healthy',
} as const;

export type DetectionSeverity = typeof DetectionSeverity[keyof typeof DetectionSeverity];


export const DetectionSeverity = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface Detection {
  id: number;
  crop: string;
  disease: string;
  diseaseHindi: string;
  type: DetectionType;
  confidence: number;
  severity: DetectionSeverity;
  description: string;
  detectedAt: string;
  /** @nullable */
  imageUrl: string | null;
  /** @nullable */
  sampleId: string | null;
}

export type RecommendationPriority = typeof RecommendationPriority[keyof typeof RecommendationPriority];


export const RecommendationPriority = {
  now: 'now',
  soon: 'soon',
  prevent: 'prevent',
} as const;

export interface Recommendation {
  title: string;
  detail: string;
  /** @nullable */
  dosage: string | null;
  priority: RecommendationPriority;
}

export interface RecommendationBundle {
  organic: Recommendation[];
  chemical: Recommendation[];
  prevention: Recommendation[];
}

export type DashboardSummaryTrendItem = {
  label: string;
  value: number;
};

export type DashboardSummaryCommonIssuesItem = {
  label: string;
  value: number;
};

export interface DashboardSummary {
  totalScans: number;
  activeFields: number;
  highRiskAlerts: number;
  cropSavedAcres: number;
  trend: DashboardSummaryTrendItem[];
  commonIssues: DashboardSummaryCommonIssuesItem[];
}

export interface FieldInput {
  name: string;
  crop: string;
  acres: number;
  village: string;
}

export type Field = FieldInput & {
  id: number;
  lastScanned: string;
};

export type AlertSeverity = typeof AlertSeverity[keyof typeof AlertSeverity];


export const AlertSeverity = {
  info: 'info',
  watch: 'watch',
  urgent: 'urgent',
} as const;

export interface Alert {
  id: number;
  title: string;
  message: string;
  severity: AlertSeverity;
  area: string;
  time: string;
  icon: string;
}

export interface CommunityPostInput {
  title: string;
  body: string;
  crop: string;
}

export type CommunityPost = CommunityPostInput & {
  id: number;
  author: string;
  replies: number;
  postedAt: string;
  expertAnswered: boolean;
};

export interface ExpertRequestInput {
  subject: string;
  message: string;
  preferredLanguage: string;
}

export type ExpertRequestStatus = typeof ExpertRequestStatus[keyof typeof ExpertRequestStatus];


export const ExpertRequestStatus = {
  queued: 'queued',
  assigned: 'assigned',
  answered: 'answered',
} as const;

export type ExpertRequest = ExpertRequestInput & {
  id: number;
  status: ExpertRequestStatus;
  eta: string;
};