# lib/api-zod/src/generated/types/alert.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * CropGuard AI API
 * OpenAPI spec version: 0.1.0
 */
import type { AlertSeverity } from './alertSeverity';

export interface Alert {
  id: number;
  title: string;
  message: string;
  severity: AlertSeverity;
  area: string;
  time: string;
  icon: string;
}