# lib/api-zod/src/generated/types/communityPostInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * CropGuard AI API
 * OpenAPI spec version: 0.1.0
 */

export interface CommunityPostInput {
  title: string;
  body: string;
  crop: string;
}