# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * CropGuard AI API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryCommonIssuesItem } from './dashboardSummaryCommonIssuesItem';
import type { DashboardSummaryTrendItem } from './dashboardSummaryTrendItem';

export interface DashboardSummary {
  totalScans: number;
  activeFields: number;
  highRiskAlerts: number;
  cropSavedAcres: number;
  trend: DashboardSummaryTrendItem[];
  commonIssues: DashboardSummaryCommonIssuesItem[];
}