# lib/api-zod/src/generated/types/detection.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * CropGuard AI API
 * OpenAPI spec version: 0.1.0
 */
import type { DetectionSeverity } from './detectionSeverity';
import type { DetectionType } from './detectionType';

export interface Detection {
  id: number;
  crop: string;
  disease: string;
  diseaseHindi: string;
  type: DetectionType;
  confidence: number;
  severity: DetectionSeverity;
  description: string;
  detectedAt: string;
  /** @nullable */
  imageUrl: string | null;
  /** @nullable */
  sampleId: string | null;
}