# lib/api-zod/src/generated/types/detectionInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * CropGuard AI API
 * OpenAPI spec version: 0.1.0
 */

export interface DetectionInput {
  sampleId?: string;
  /** @nullable */
  imageDataUrl?: string | null;
  crop: string;
  /** @nullable */
  fieldId?: number | null;
}