# lib/api-zod/src/generated/types/expertRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * CropGuard AI API
 * OpenAPI spec version: 0.1.0
 */
import type { ExpertRequestInput } from './expertRequestInput';
import type { ExpertRequestStatus } from './expertRequestStatus';

export type ExpertRequest = ExpertRequestInput & {
  id: number;
  status: ExpertRequestStatus;
  eta: string;
};