# lib/api-zod/src/generated/types/fieldInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * CropGuard AI API
 * OpenAPI spec version: 0.1.0
 */

export interface FieldInput {
  name: string;
  crop: string;
  acres: number;
  village: string;
}