# lib/api-zod/src/generated/types/recommendation.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * CropGuard AI API
 * OpenAPI spec version: 0.1.0
 */
import type { RecommendationPriority } from './recommendationPriority';

export interface Recommendation {
  title: string;
  detail: string;
  /** @nullable */
  dosage: string | null;
  priority: RecommendationPriority;
}