# lib/api-zod/src/generated/types/recommendationBundle.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * CropGuard AI API
 * OpenAPI spec version: 0.1.0
 */
import type { Recommendation } from './recommendation';

export interface RecommendationBundle {
  organic: Recommendation[];
  chemical: Recommendation[];
  prevention: Recommendation[];
}